Add files via upload
This exercise uses an array, a loop, and generates a string of HTML. Start with an array of strings. For each element in the array, generate an HTML paragraph. At the beginning of each paragraph, include a sequential number that numbers the paragraphs from 1 to the number of elements in the array. Concatenate all the paragraphs together into one string of HTML, then display the string.