Add files via upload
This exercise uses an array, a loop, and generates a string of HTML. Start with an array of strings. Generate the HTML for a dropdown list (<select> element with nested <option> elements). Each element in the array corresponds to an option in the dropdown list. Do not include a value attribute for each option. Then display the dropdown list.