Add files via upload
This exercise uses an array, a loop, and generates a string of HTML. Start with an array of strings. Generate the HTML for a dropdown list (<select> element with nested <option> elements). Each element in the array corresponds to an option in the dropdown list. The value attribute for each option should be the array element's index. Then display the dropdown list.