Add files via upload
This exercise generates <img> tags. Start with an array of image filenames and assume the images are stored in an "images" folder. A second array contains the alternate text for the images. Write a function that takes both arrays as arguments and returns a string of HTML containing the generated <img> tags. Each element in the arrays corresponds to an <img> tag. Separate each <img> tag with a line break.