Add files via upload
This exercise generates radio buttons. Start with an array of text labels, one for each radio button. A second array contains the values for the radio buttons. Write a function that takes both arrays as arguments and returns a string of HTML containing the generated radio buttons. Each element in the arrays corresponds to a radio button. Separate each radio button with a line break.

Modify the previous exercise by replacing the 2 arrays with 1 associative array.